# [reader] optimize hot Reader.Read() path